Criando modelo para Video
#Close137
