+++ Pipfile
@@ -8,6 +8,7 @@ flake8 = "*"
 pytest-django = "*"
 pytest-cov = "*"
 codecov = "*"
+model-mommy = "*"
 
 [packages]
 django = "*"
